pyproject.toml
[tool.poetry]
name = "media-server"
version = "0.1.0"
description = ""
authors = ["Alexey Iskovskikh <iskovskikh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^71.0.3"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
punq = "^0.7.0"
pydantic-settings = "^2.3.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
